Reworking of code editor to support read-only scripts, added combat engage sound

+++ Source/Engine/Hackables/CodeEditor/ScriptEntry.cpp
@@ -12,26 +12,27 @@
 
 using namespace cocos2d;
 
-ScriptEntry* ScriptEntry::create(ConstantString* scriptName, std::string script, std::function<void(ScriptEntry*)> onScriptEntryClick, std::function<void(ScriptEntry*)> onDeleteClick)
+ScriptEntry* ScriptEntry::create(LocalizedString* scriptName, std::string script, bool isReadOnly, std::function<void(ScriptEntry*)> onScriptEntryClick, std::function<void(ScriptEntry*)> onDeleteClick)
 {
-	ScriptEntry* instance = new ScriptEntry(scriptName, script, onScriptEntryClick, onDeleteClick);
+	ScriptEntry* instance = new ScriptEntry(scriptName, script, isReadOnly, onScriptEntryClick, onDeleteClick);
 
 	instance->autorelease();
 
 	return instance;
 }
 
-ScriptEntry::ScriptEntry(ConstantString* scriptName, std::string script, std::function<void(ScriptEntry*)> onScriptEntryClick, std::function<void(ScriptEntry*)> onDeleteClick)
+ScriptEntry::ScriptEntry(LocalizedString* scriptName, std::string script, bool isReadOnly, std::function<void(ScriptEntry*)> onScriptEntryClick, std::function<void(ScriptEntry*)> onDeleteClick)
 {
 	this->scriptName = scriptName;
 	this->script = script;
+	this->isReadOnly = isReadOnly;
 	this->onScriptEntryClick = onScriptEntryClick;
 	this->onDeleteClick = onDeleteClick;
+	this->deleteButton = nullptr;
 
 	this->backPlate = ClickableNode::create(UIResources::Menus_HackerModeMenu_ScriptEntry, UIResources::Menus_HackerModeMenu_ScriptEntrySelected);
 	this->selectedSprite = Sprite::create(UIResources::Menus_HackerModeMenu_SelectedScriptArrow);
 	this->label = LocalizedLabel::create(LocalizedLabel::FontStyle::Main, LocalizedLabel::FontSize::P, this->scriptName);
-	this->deleteButton = ClickableNode::create(UIResources::Menus_HackerModeMenu_TrashCan, UIResources::Menus_HackerModeMenu_TrashCanSelected);
 
 	this->selectedSprite->setColor(Color3B::YELLOW);
 	this->selectedSprite->setAnchorPoint(Vec2(0.0f, 0.5f));
@@ -40,7 +41,12 @@ ScriptEntry::ScriptEntry(ConstantString* scriptName, std::string script, std::fu
 	this->addChild(this->backPlate);
 	this->addChild(this->selectedSprite);
 	this->addChild(this->label);
-	this->addChild(this->deleteButton);
+
+	if (!this->isReadOnly)
+	{
+		this->deleteButton = ClickableNode::create(UIResources::Menus_HackerModeMenu_TrashCan, UIResources::Menus_HackerModeMenu_TrashCanSelected);
+		this->addChild(this->deleteButton);
+	}
 }
 
 void ScriptEntry::initializePositions()
@@ -51,7 +57,11 @@ void ScriptEntry::initializePositions()
 
 	this->label->setPositionX(-this->backPlate->getContentSize().width / 2.0f + margin);
 	this->selectedSprite->setPositionX(-this->backPlate->getContentSize().width / 2.0f + margin);
-	this->deleteButton->setPositionX(this->backPlate->getContentSize().width / 2.0f - this->deleteButton->getContentSize().width - 8.0f);
+
+	if (this->deleteButton != nullptr)
+	{
+		this->deleteButton->setPositionX(this->backPlate->getContentSize().width / 2.0f - this->deleteButton->getContentSize().width - 8.0f);
+	}
 }
 
 void ScriptEntry::initializeListeners()
@@ -59,21 +69,25 @@ void ScriptEntry::initializeListeners()
 	super::initializeListeners();
 
 	this->backPlate->setMouseClickCallback([=](InputEvents::MouseEventArgs*) { this->onScriptEntryClick(this); });
-	this->deleteButton->setMouseClickCallback([=](InputEvents::MouseEventArgs*) { this->onDeleteClick(this); });
+	
+	if (this->deleteButton != nullptr)
+	{
+		this->deleteButton->setMouseClickCallback([=](InputEvents::MouseEventArgs*) { this->onDeleteClick(this); });
+	}
 }
 
 void ScriptEntry::toggleSelected(bool isSelected)
 {
-	const float margin = 16.0f;
-	const float selectedOffset = 32.0f;
+	const float Margin = 16.0f;
+	const float SelectedOffset = 32.0f;
 
 	this->selectedSprite->setVisible(isSelected);
 
 	this->label->setTextColor(isSelected ? Color4B::YELLOW : Color4B::WHITE);
-	this->label->setPositionX(-this->backPlate->getContentSize().width / 2.0f + margin + (isSelected ? selectedOffset : 0.0f));
+	this->label->setPositionX(-this->backPlate->getContentSize().width / 2.0f + Margin + (isSelected ? SelectedOffset : 0.0f));
 }
 
-ConstantString* ScriptEntry::getName()
+LocalizedString* ScriptEntry::getName()
 {
 	return this->scriptName;
 }
